Added storing gotten pokemon to MongoDB
