+++ src/main/java/com/qa/pokeAPI/repository/PokemonRepository.java
@@ -1,0 +1,14 @@
+package com.qa.pokeAPI.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.qa.pokeAPI.domain.PokemonData;
+
+public interface PokemonRepository extends MongoRepository<PokemonData, String>
+{
+
+	public PokemonData findByName(String name);
+
+	public PokemonData findByPokeId(int pokeId);
+
+}
